Todos Provider component logic was implemented
