+++ src/index.tsx
@@ -5,7 +5,12 @@ import './styles/todo-list.css';
 import './styles/filters.css';
 
 import { App } from './App';
+import { TodosProvider } from './Context/TodosProvider';
 
 const container = document.getElementById('root') as HTMLDivElement;
 
-createRoot(container).render(<App />);
+createRoot(container).render(
+  <TodosProvider>
+    <App />
+  </TodosProvider>,
+);
